Refactor datasets
